Use current version of the provider interface
There are currently two versions of the provider interface the current
versioned interface [1] and the legacy unversioned interface [2]. The
versioned interface has abstract base classes named ProviderV1,
BackendV1, etc where the version of each abstract interface is
incremeneted as the API is modified to maintain backwards compatibility.
This interface provides a simpler experience for users and looks like
what the ionq provider was intended to be used with. The legacy providers
interface has classes named BaseProvider, BaseBackend, etc and existed
before the versioned interface was added in terra 0.16.0. This is in a
state of pre-deprecation because not everything has had a chance to
migrate to it yet. But all new providers should be using the versioned
interface if possible. This commit updates the ionq provider to use the
current version of the interface.

[1] https://qiskit.org/documentation/apidoc/providers.html
[2] https://qiskit.org/documentation/apidoc/providers.html#legacy-provider-interface-base-objects-qiskit-providers