Correctly use options object
In #43 the ionq backend object was updated to use BackendV1 and handle
it's user configurable options/settings via an Options object. The goal
of that object is to allow for runtime setting of defaults for all the
user tunable options around running a circuit. The _default_options
class method should return an options object with all the available
options for a backend. The values set in the options object (which can
be changed by users prior to calling run()) should be used as the default
unless overridden by a kwarg with the same name to run(). This commit
fixes the behavior so we're using the options value. It also raises a
warning for users if an unsupported option is used so they know that the
kwarg being passed to run doesn't have any effect.