Change overriden method in IonQGateError to __repr__
This was overriding the message defined in the class, so that users would see
IonQGateError(gate_name='foo')
instead of
IonQGateError('gate 'foo' not supported on IonQ backends. Please use the qiskit.transpile method or manually rewrite to remove the gate')

This way, developers will see the former and users the latter, as was originally intended.