Remove inverses
Under the advisement of a distinguished physicist peer who did not want me to embarass myself and said
'There is a phase issue, which in almost all case won't manifest itself.  But it's related to what people say as 'in the world of a spin-1/2 partical, 2pi rotation doesn't brings you back to origin''
I admittedly don't understand that bit at all, and promptly agreed to remove the inverses.

Test updated to check unitarity